==> models/saiva-3-day-hosp-v3/src/notebooks/00-setup.ipynb <==
# This notebook is responsible for the initial setup - clean all files from previous runs, load client configurations and save them

import sys
import pandas as pd
import shutil
import os
import glob

from omegaconf import OmegaConf

from shared.constants import saiva_api, LOCAL_TRAINING_CONFIG_PATH
from shared.utils import load_config

from eliot import to_file
to_file(sys.stdout)

### =============== Delete all files related to previous model run/training ==================

# file_types = ['*.pickle', '*.json', '*.csv', '*.png', '*.txt']

# for file_type in file_types:
#     files = glob.glob(file_type)
#     for file in files:
#         print(f"Removed {file}")
#         os.remove(file)

### =============== Delete all files inside /data/processed & /data/raw ==================

# import shutil
# from pathlib import Path

# processed_path = Path('/data/processed')
# raw_path = Path('/data/raw')
# shutil.rmtree(processed_path)
# shutil.rmtree(raw_path)
# print("Success.......")

### ====================== Clean Configuration ======================

shutil.rmtree(f'{LOCAL_TRAINING_CONFIG_PATH}generated', ignore_errors=True)
os.makedirs(f'{LOCAL_TRAINING_CONFIG_PATH}generated', exist_ok=True)

## Load config

config = load_config(LOCAL_TRAINING_CONFIG_PATH)
training_config = config.training_config

print("TRAIN_START_DATE:", training_config.training_metadata.experiment_dates.train_start_date)
print("TEST_END_DATE:", training_config.training_metadata.experiment_dates.test_end_date)
print("CLIENTS:", [organization_config.organization_id for organization_config in training_config.organization_configs_setup])

### ====================== Load ml model organization configs ======================

ml_model_organization_configs = []

for organization_config in training_config.organization_configs_setup:
    ml_model_organization_configs.append(
        saiva_api.organization_ml_model_configs.get(
            org_id=organization_config.organization_id, 
            ml_model_org_config_id=organization_config.ml_model_org_config_id
        )
    )

### The configuration will may point to the prod databases - run the next cell to point it to dev if you want to fetch the data + override existing models for local training purposes

for ml_model_organization_config in ml_model_organization_configs:
    ml_model_organization_config.source_database_credentials_secret_id = 'dev-sqlserver'
    for i, facility_config in enumerate(ml_model_organization_config.facilities):
        ml_model_organization_config.facilities[i] = facility_config.__dict__
        ml_model_organization_config.facilities[i]['ml_model'] = None
        ml_model_organization_config.facilities[i]['trained_model'] = None

### Store the configuration to be able to change the configuration manually

conf = OmegaConf.create({'training_config': {'ml_model_org_configs': [config.__dict__ for config in ml_model_organization_configs]}})
OmegaConf.save(conf, f'{LOCAL_TRAINING_CONFIG_PATH}generated/ml_model_organization_configs.yaml')
